test(fixture): Start dep upgrade fixture
We're limited by the fact that we have everything in the workspace and
can only have one package with a given name in the workspace.

As we solidify our strategy, maybe we can drop these being in the
workspace and get this going.  In the mean time, this will at least giv
us helpful data.