Inherit stderr when running `cargo doc` to print progress
Currently, when running `cargo api` on a big crate, there is no output for a
long time, since `cargo doc` runs silently.

Inherit stderr to print the progress of running `cargo doc`. We don't want to
inherit stdout since it should be easy to pipe the output of `cargo api` to a
file without progress noise. Luckly, `cargo` prints progress to stderr.